requirements.txt
streamlit>=1.31.0
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10
faiss-cpu>=1.7.4
tiktoken>=0.5.2
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pypdf>=3.0.0
duckduckgo-search>=4.1.1